fix: improved routing and animation handling logic
