+++ src/script.js
@@ -1,5 +1,3 @@
-import { route } from "./utils/router.js";
-
 function webgl_support() {
   //https://stackoverflow.com/questions/11871077/proper-way-to-detect-webgl-support
   try {
@@ -14,7 +12,6 @@ function webgl_support() {
   }
 }
 document.addEventListener("DOMContentLoaded", () => {
-  route();
   if (!webgl_support()) {
     console.error(
       "WebGL is not supported on your browser. Please visit https://get.webgl.org/webgl2/ for more."
