+++ src/utils/router.js
@@ -1,39 +1,15 @@
 import Navigo from "navigo";
-import { computer, renderPage } from "../Computer.js";
+import { renderPage } from "../Computer.js";
 
-let isModelLoaded = false;
-let scene, camera, renderer;
+const router = new Navigo("/", { hash: true });
 
-export function route() {
-  if (!isModelLoaded) {
-    const result = computer();
-    scene = result.scene;
-    camera = result.camera;
-    renderer = result.renderer;
-    isModelLoaded = true;
-  }
-
-  // eslint-disable-next-line no-undef
-  const router = new Navigo("/", { hash: true });
-
+export function setupRouter(scene, camera, renderer) {
   router
-    .on("/", () => {
-      renderPage(scene, camera, renderer, "/");
-    })
-    .on("/about", () => {
-      renderPage(scene, camera, renderer, "/about");
-    })
-    .on("/projects", () => {
-      renderPage(scene, camera, renderer, "/projects");
-    })
-    .on("/experience", () => {
-      renderPage(scene, camera, renderer, "/experience");
-    })
-    .on("/contact", () => {
-      renderPage(scene, camera, renderer, "/contact");
-    })
-    .resolve();
-
-  router.updatePageLinks();
-  return router;
+    .on("/", () => renderPage(scene, camera, renderer, "/"))
+    .on("/about", () => renderPage(scene, camera, renderer, "/about"))
+    .on("/projects", () => renderPage(scene, camera, renderer, "/projects"))
+    .on("/experience", () => renderPage(scene, camera, renderer, "/experience"))
+    .on("/contact", () => renderPage(scene, camera, renderer, "/contact"));
 }
+
+export { router };
